rename file to match exported type inside
