+++ src/common/types/store-data/details-view-store-data.ts
