+++ src/tests/unit/common/details-view-store-data-builder.ts
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 import { DetailsViewStore } from 'background/stores/details-view-store';
-import { DetailsViewStoreData } from '../../../common/types/store-data/details-view-data';
+import { DetailsViewStoreData } from '../../../common/types/store-data/details-view-store-data';
 import { DetailsViewRightContentPanelType } from '../../../DetailsView/components/left-nav/details-view-right-content-panel-type';
 import { BaseDataBuilder } from './base-data-builder';
