+++ src/tests/unit/tests/DetailsView/store-mocks.ts
@@ -23,7 +23,7 @@ import { UnifiedScanResultStore } from '../../../../background/stores/unified-sc
 import { FeatureFlags } from '../../../../common/feature-flags';
 import { AssessmentStoreData } from '../../../../common/types/store-data/assessment-result-data';
 import { CommandStoreData } from '../../../../common/types/store-data/command-store-data';
-import { DetailsViewStoreData } from '../../../../common/types/store-data/details-view-data';
+import { DetailsViewStoreData } from '../../../../common/types/store-data/details-view-store-data';
 import { FeatureFlagStoreData } from '../../../../common/types/store-data/feature-flag-store-data';
 import { LaunchPanelStoreData } from '../../../../common/types/store-data/launch-panel-store-data';
 import { ScopingStoreData } from '../../../../common/types/store-data/scoping-store-data';
